Rename directory name from wraper to wrapper
