+++ bootstrap/init.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/pingcap/tidb-foresight/wraper/db"
+	"github.com/pingcap/tidb-foresight/wrapper/db"
 	log "github.com/sirupsen/logrus"
 )
