+++ model/logs/model.go
@@ -1,7 +1,7 @@
 package logs
 
 import (
-	"github.com/pingcap/tidb-foresight/wraper/db"
+	"github.com/pingcap/tidb-foresight/wrapper/db"
 )
 
 type Model interface {
