+++ model/report/model.go
@@ -2,7 +2,7 @@ package report
 
 import (
 	"github.com/pingcap/tidb-foresight/utils"
-	"github.com/pingcap/tidb-foresight/wraper/db"
+	"github.com/pingcap/tidb-foresight/wrapper/db"
 )
 
 type Model interface {
