+++ server/server.go
@@ -13,7 +13,7 @@ import (
 	"github.com/pingcap/tidb-foresight/log/searcher"
 	"github.com/pingcap/tidb-foresight/model"
 	"github.com/pingcap/tidb-foresight/utils"
-	"github.com/pingcap/tidb-foresight/wraper/db"
+	"github.com/pingcap/tidb-foresight/wrapper/db"
 	log "github.com/sirupsen/logrus"
 )
